Move _ChildProtocolWrapper to smartprotocol.py
