+++ kasa/tests/test_protocol.py
@@ -482,6 +482,9 @@ def _get_subclasses(of_class):
     "class_name_obj", _get_subclasses(BaseProtocol), ids=lambda t: t[0]
 )
 def test_protocol_init_signature(class_name_obj):
+    if class_name_obj[0].startswith("_"):
+        pytest.skip("Skipping internal protocols")
+        return
     params = list(inspect.signature(class_name_obj[1].__init__).parameters.values())
 
     assert len(params) == 2
